requirements.txt
pytorch
pyyaml
colorlog
